pythonProject/requirements.txt
pytest
request
pydantic
typing
re
uvicorn
fastapi